requirements.txt
Django
djangorestframework
django-simple-captcha==0.5.13
django-simple-menu
markdown
future
